Add a GUI for the Duke Application Add JavaFX for duke application todo: edit the linebreaks and wrappers to get a better display. TODO: Add initial message and username setting
