+++ src/main/java/gui/Launcher.java
@@ -1,0 +1,13 @@
+package gui;
+
+import gui.ui.MainLauncher;
+import javafx.application.Application;
+
+/**
+ * Runs application launcher for GUI
+ */
+public class Launcher {
+    public static void main(String[] args) {
+        Application.launch(MainLauncher.class, args);
+    }
+}
