+++ text-ui-test/EXPECTED.TXT
@@ -7,7 +7,7 @@ Hello from
 
 	_________________________________________________
 
-	Hello! I'm Duke
+	Hello! I'm duke.Duke
 	What can I do for you
 
 	_________________________________________________
@@ -15,24 +15,24 @@ Hello from
 ####################################################################################################
 	Got it. I've added this task:
 	  [T][✘]  taskAbc 1231
-	Now you have 1 tasks in the list.
+	Now you have 1 duke.tasks in the list.
 ####################################################################################################
 
 ####################################################################################################
-	Here are the tasks in your list:
+	Here are the duke.tasks in your list:
 	1. [T][✘]  taskAbc 1231
 ####################################################################################################
 
 ####################################################################################################
 	Got it. I've added this task:
 	  [D][✘] null (by: null)
-	Now you have 2 tasks in the list.
+	Now you have 2 duke.tasks in the list.
 ####################################################################################################
 
 ####################################################################################################
 	Got it. I've added this task:
 	  [D][✘]  Say Hi  (by:  tofu)
-	Now you have 3 tasks in the list.
+	Now you have 3 duke.tasks in the list.
 ####################################################################################################
 
 ####################################################################################################
@@ -62,7 +62,7 @@ Hello from
 ####################################################################################################
 
 ####################################################################################################
-	Here are the tasks in your list:
+	Here are the duke.tasks in your list:
 	1. [T][✓]  taskAbc 1231
 	2. [D][✓] null (by: null)
 	3. [D][✓]  Say Hi  (by:  tofu)
@@ -71,7 +71,7 @@ Hello from
 ####################################################################################################
 	Got it. I've added this task:
 	  [E][✘]  hello friends /by  (at:  TIME)
-	Now you have 4 tasks in the list.
+	Now you have 4 duke.tasks in the list.
 ####################################################################################################
 
 	_________________________________________________
